requirements/local.txt
# Packages required for local development
-r test.txt

django-debug-toolbar==1.8
nose==1.3.7
pudb==2018.1
